+++ src/components/Modal/Modal.tsx
@@ -2,18 +2,18 @@ import classNames from 'classnames';
 import React, { memo } from 'react';
 
 import './Modal.scss';
-import { useAppDispatch, useAppSelector } from '../../app/hooks';
-import { actions } from '../../features/modal/itActive';
+
+import { useAppDispatch } from '../../hooks/useAppDispatch';
+import { useAppSelector } from '../../hooks/useAppSelector';
+import { actions } from '../../store/actions/modalActions';
 
 type PropTypes = {
   children: React.ReactNode;
 };
 
-export const Modal: React.FC<PropTypes> = memo(({
-  children,
-}) => {
+export const Modal: React.FC<PropTypes> = memo(({ children }) => {
   const dispatch = useAppDispatch();
-  const { isActive } = useAppSelector(state => state.isActive);
+  const { isActive } = useAppSelector((state) => state.isActive);
 
   const handleClick = () => {
     dispatch(actions.setIsActive(false));
@@ -22,18 +22,15 @@ export const Modal: React.FC<PropTypes> = memo(({
   return (
     <button
       type="button"
-      className={classNames(
-        'modal',
-        {
-          'modal--active': isActive,
-        },
-      )}
+      className={classNames('modal', {
+        'modal--active': isActive,
+      })}
       onClick={handleClick}
     >
       <button
         type="button"
         className="modal__content"
-        onClick={e => e.stopPropagation()}
+        onClick={(e) => e.stopPropagation()}
       >
         {children}
       </button>
