+++ src/components/TodoList/TodoList.tsx
@@ -1,10 +1,10 @@
-import { useAppSelector } from '../../app/hooks';
+import { useAppSelector } from '../../hooks/useAppSelector';
 import { Todo } from '../Todo';
 
 import './TodoList.scss';
 
 export const TodoList = () => {
-  const { todos } = useAppSelector(state => state.todos);
+  const { todos } = useAppSelector((state) => state.todos);
 
   if (!todos.length) {
     return <p className="todolist__empty">There is nothing left</p>;
@@ -12,7 +12,7 @@ export const TodoList = () => {
 
   return (
     <ul className="todolist">
-      {todos.map(todo => (
+      {todos.map((todo) => (
         <Todo key={todo.id} todo={todo} />
       ))}
     </ul>
