+++ src/hooks/useAppDispatch.ts
@@ -1,0 +1,5 @@
+import { useDispatch } from 'react-redux';
+
+import { AppDispatch } from '../store/store';
+
+export const useAppDispatch: () => AppDispatch = useDispatch;
