+++ src/hooks/useAppSelector.ts
@@ -1,0 +1,5 @@
+import { TypedUseSelectorHook, useSelector } from 'react-redux';
+
+import { RootState } from '../store/store';
+
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
