+++ src/store/actions/modalActions.ts
@@ -1,0 +1,11 @@
+export type IsActiveAction = {
+  type: 'status/change',
+  payload: boolean;
+};
+
+const setIsActive = (isActive: boolean): IsActiveAction => ({
+  type: 'status/change',
+  payload: isActive,
+});
+
+export const actions = { setIsActive };
