+++ src/store/actions/queryAction.ts
@@ -1,0 +1,11 @@
+export type QueryAction = {
+  type: 'query/change',
+  payload: string;
+};
+
+const setQuery = (query: string): QueryAction => ({
+  type: 'query/change',
+  payload: query,
+});
+
+export const actions = { setQuery };
