+++ src/store/actions/todosActions.ts
@@ -1,0 +1,13 @@
+import { ITodo } from '../../types/Todo';
+
+export type AddTodoAction = {
+  type: 'todo/add';
+  payload: ITodo;
+};
+
+const addTodo = (todo: ITodo): AddTodoAction => ({
+  type: 'todo/add',
+  payload: todo,
+});
+
+export const actions = { addTodo };
