+++ src/store/reducers/queryReducer.ts
@@ -1,3 +1,5 @@
+import { QueryAction } from '../actions/queryAction';
+
 type QueryState = {
   query: string,
 };
@@ -5,18 +7,6 @@ type QueryState = {
 const initialState: QueryState = {
   query: '',
 };
-
-type QueryAction = {
-  type: 'query/change',
-  payload: string;
-};
-
-const setQuery = (query: string): QueryAction => ({
-  type: 'query/change',
-  payload: query,
-});
-
-export const actions = { setQuery };
 
 type Action = QueryAction;
